Crate constraints

The symbolic algebra system.

Modules

ops
Expression operations.

Structs

Equation

SystemOfEquations
A builder for constructing a system of equations and solving them.

Enums

BinaryOperation
An operation that can be applied to two arguments.

Expression
An expression.

Parameter
A free variable.

ParseError
Possible errors that may occur while parsing.

TokenKind
The kinds of token that can appear in an Expression's text form.

Functions

parse
Parse an Expression tree from some text.