constraints

Enum SelectedTab

pub(crate) enum SelectedTab {
    Min,
    Max,
    Length,
    Percentage,
    Ratio,
    Fill,
}

Tabs for the different examples

The order of the variants is the order in which they are displayed.

Variants

Min

Max

Length

Percentage

Ratio

Fill

Implementations

impl SelectedTab

pub(crate) const fn from_repr(discriminant: usize) -> Option<SelectedTab>

Try to create Self from the raw representation

impl SelectedTab

pub(crate) fn previous(self) -> Self

Get the previous tab, if there is no previous tab return the current tab.

pub(crate) fn next(self) -> Self

Get the next tab, if there is no next tab return the current tab.

pub(crate) const fn get_example_count(self) -> u16

pub(crate) fn to_tab_title(value: Self) -> Line<'static>

impl SelectedTab

pub(crate) fn render_length_example(area: Rect, buf: &mut Buffer)

pub(crate) fn render_percentage_example(area: Rect, buf: &mut Buffer)

pub(crate) fn render_ratio_example(area: Rect, buf: &mut Buffer)

pub(crate) fn render_fill_example(area: Rect, buf: &mut Buffer)

pub(crate) fn render_min_example(area: Rect, buf: &mut Buffer)

pub(crate) fn render_max_example(area: Rect, buf: &mut Buffer)