Module constraints::ops

Expression operations.

Structs

Builtins
The set of builtin functions.

Enums

EvaluationError
Errors that may occur while evaulating an Expression.

Traits

Context
Contextual information used when evaluating an Expression.

Functions

evaluate

fold_constants
Simplify an expression by evaluating all constant operations.

partial_derivative
Calculate an Expression's partial derivative with respect to a particular Parameter.

substitute
Replace all references to a Parameter with an Expression.